Convert dashboard, repositories, and search pages to Server Components
Extract interactive content into dedicated Client Component files
(_components/dashboard-content.tsx, repositories-content.tsx,
search-content.tsx) and make the page.tsx files thin Server Component
shells. This enables static prerendering of the page shell with
client-side hydration for interactive elements, and allows the
loading.tsx skeletons added earlier to stream properly.
